requirements.txt
python-telegram-bot==20.7
reportlab==4.1.0
python-telegram-bot[job-queue]==20.7
psycopg2-binary==2.9.9